Rm unused derive impls (#17)
* Rm unused derive impls

Remove unused impls from various types.

I would not impl Debug or Clone by without need.
I would also not impl Eq without a need.

In the case of Eq, be careful implementing it on complex types.
What does equlaity mean for a CDR?
Equality for such a complex object is context dependant.

* Rm copy impls on large structs

Be careful implementing Copy on structs larger than the machine word size. This can lead to a lot of silent copying when a reference (one word) does the job.

The resulting changes to the code make it clear where data is being cloned.